backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
gunicorn==20.1.0
requests==2.28.2
python-dotenv==1.0.0
pymongo[srv]==4.5.0
Werkzeug==2.2.3
groq==0.4.1